requirements.txt
# Core dependencies
web3==6.11.1
python-telegram-bot==13.7
python-dotenv==1.0.0

# Analysis and utilities
pandas==2.1.3
numpy==1.24.3

# Development tools
black==23.11.0
pylint==3.0.2
pytest==7.4.3
